+++ src/commands/dependencies/unedit.ts
@@ -30,10 +30,10 @@ export async function uneditDependency(
 ) {
     const currentFolder = folder ?? ctx.currentFolder;
     if (!currentFolder) {
-        ctx.outputChannel.log("currentFolder is not set.");
+        ctx.logger.debug("currentFolder is not set.");
         return false;
     }
-    ctx.outputChannel.log(`unedit dependency ${identifier}`, currentFolder.name);
+    ctx.logger.debug(`unedit dependency ${identifier}`, currentFolder.name);
     const status = `Reverting edited dependency ${identifier} (${currentFolder.name})`;
     return await ctx.statusItem.showStatusWhileRunning(status, async () => {
         return await uneditFolderDependency(currentFolder, identifier, ctx);
@@ -89,12 +89,12 @@ async function uneditFolderDependency(
             );
 
             if (result === "No") {
-                ctx.outputChannel.log(execError.stderr, folder.name);
+                ctx.logger.error(execError.stderr, folder.name);
                 return false;
             }
             await uneditFolderDependency(folder, identifier, ctx, ["--force"]);
         } else {
-            ctx.outputChannel.log(execError.stderr, folder.name);
+            ctx.logger.error(execError.stderr, folder.name);
             void vscode.window.showErrorMessage(`${execError.stderr}`);
         }
         return false;