+++ src/toolchain/toolchain.ts
@@ -18,7 +18,6 @@ import * as os from "os";
 import * as plist from "plist";
 import * as vscode from "vscode";
 import configuration from "../configuration";
-import { SwiftOutputChannel } from "../ui/SwiftOutputChannel";
 import { execFile, execSwift } from "../utilities/utilities";
 import { expandFilePathTilde, fileExists, pathExists } from "../utilities/filesystem";
 import { Version } from "../utilities/version";
@@ -26,6 +25,7 @@ import { BuildFlags } from "./BuildFlags";
 import { Sanitizer } from "./Sanitizer";
 import { lineBreakRegex } from "../utilities/tasks";
 import { Swiftly } from "./swiftly";
+import { SwiftLogger } from "../logging/SwiftLogger";
 /**
  * Contents of **Info.plist** on Windows.
  */
@@ -118,15 +118,12 @@ export class SwiftToolchain {
         this.swiftVersionString = targetInfo.compilerVersion;
     }
 
-    static async create(
-        folder?: vscode.Uri,
-        outputChannel?: vscode.OutputChannel
-    ): Promise<SwiftToolchain> {
+    static async create(folder?: vscode.Uri, logger?: SwiftLogger): Promise<SwiftToolchain> {
         const { path: swiftFolderPath, isSwiftlyManaged } = await this.getSwiftFolderPath(
             folder,
-            outputChannel
+            logger
         );
-        const toolchainPath = await this.getToolchainPath(swiftFolderPath, folder, outputChannel);
+        const toolchainPath = await this.getToolchainPath(swiftFolderPath, folder, logger);
         const targetInfo = await this.getSwiftTargetInfo(
             this._getToolchainExecutable(toolchainPath, "swift")
         );
@@ -142,13 +139,15 @@ export class SwiftToolchain {
                 swiftFolderPath,
                 swiftVersion,
                 runtimePath,
-                customSDK ?? defaultSDK
+                customSDK ?? defaultSDK,
+                logger
             ),
             this.getSwiftTestingPath(
                 targetInfo,
                 swiftVersion,
                 runtimePath,
-                customSDK ?? defaultSDK
+                customSDK ?? defaultSDK,
+                logger
             ),
             this.getSwiftPMTestingHelperPath(toolchainPath),
         ]);
@@ -516,13 +515,13 @@ export class SwiftToolchain {
         return str;
     }
 
-    logDiagnostics(channel: SwiftOutputChannel) {
-        channel.logDiagnostic(this.diagnostics);
+    logDiagnostics(logger: SwiftLogger) {
+        logger.debug(this.diagnostics);
     }
 
     private static async getSwiftFolderPath(
         cwd?: vscode.Uri,
-        outputChannel?: vscode.OutputChannel
+        logger?: SwiftLogger
     ): Promise<{ path: string; isSwiftlyManaged: boolean }> {
         try {
             let swift: string;
@@ -588,7 +587,7 @@ export class SwiftToolchain {
                 isSwiftlyManaged,
             };
         } catch (error) {
-            outputChannel?.appendLine(`Failed to find swift executable: ${error}`);
+            logger?.error(`Failed to find swift executable: ${error}`);
             throw Error("Failed to find swift executable");
         }
     }
@@ -624,7 +623,7 @@ export class SwiftToolchain {
     private static async getToolchainPath(
         swiftPath: string,
         cwd?: vscode.Uri,
-        channel?: vscode.OutputChannel
+        logger?: SwiftLogger
     ): Promise<string> {
         try {
             switch (process.platform) {
@@ -645,7 +644,7 @@ export class SwiftToolchain {
                         return path.dirname(configuration.path);
                     }
 
-                    const swiftlyToolchainLocation = await Swiftly.toolchain(channel, cwd);
+                    const swiftlyToolchainLocation = await Swiftly.toolchain(logger, cwd);
                     if (swiftlyToolchainLocation) {
                         return swiftlyToolchainLocation;
                     }
@@ -749,7 +748,8 @@ export class SwiftToolchain {
         targetInfo: SwiftTargetInfo,
         swiftVersion: Version,
         runtimePath: string | undefined,
-        sdkroot: string | undefined
+        sdkroot: string | undefined,
+        logger?: SwiftLogger
     ): Promise<string | undefined> {
         if (process.platform !== "win32") {
             return undefined;
@@ -759,7 +759,8 @@ export class SwiftToolchain {
             targetInfo,
             swiftVersion,
             runtimePath,
-            sdkroot
+            sdkroot,
+            logger
         );
     }
 
@@ -775,7 +776,8 @@ export class SwiftToolchain {
         swiftFolderPath: string,
         swiftVersion: Version,
         runtimePath: string | undefined,
-        sdkroot: string | undefined
+        sdkroot: string | undefined,
+        logger?: SwiftLogger
     ): Promise<string | undefined> {
         switch (process.platform) {
             case "darwin": {
@@ -793,7 +795,8 @@ export class SwiftToolchain {
                     targetInfo,
                     swiftVersion,
                     runtimePath,
-                    sdkroot
+                    sdkroot,
+                    logger
                 );
             }
         }
@@ -805,7 +808,8 @@ export class SwiftToolchain {
         targetInfo: SwiftTargetInfo,
         swiftVersion: Version,
         runtimePath: string | undefined,
-        sdkroot: string | undefined
+        sdkroot: string | undefined,
+        logger?: SwiftLogger
     ): Promise<string | undefined> {
         // look up runtime library directory for XCTest/Testing alternatively
         const fallbackPath =
@@ -838,9 +842,7 @@ export class SwiftToolchain {
         const plistKey = type === "XCTest" ? "XCTEST_VERSION" : "SWIFT_TESTING_VERSION";
         const version = infoPlist.DefaultProperties[plistKey];
         if (!version) {
-            new SwiftOutputChannel("swift").appendLine(
-                `Warning: ${platformManifest} is missing the ${plistKey} key.`
-            );
+            logger?.warn(`${platformManifest} is missing the ${plistKey} key.`);
             return undefined;
         }