+++ src/ui/ToolchainSelection.ts
@@ -19,6 +19,7 @@ import { SwiftToolchain } from "../toolchain/toolchain";
 import configuration from "../configuration";
 import { Commands } from "../commands";
 import { Swiftly } from "../toolchain/swiftly";
+import { SwiftLogger } from "../logging/SwiftLogger";
 
 /**
  * Open the installation page on Swift.org
@@ -154,6 +155,7 @@ type SelectToolchainItem = SwiftToolchainItem | ActionItem | SeparatorItem;
  */
 async function getQuickPickItems(
     activeToolchain: SwiftToolchain | undefined,
+    logger: SwiftLogger,
     cwd?: vscode.Uri
 ): Promise<SelectToolchainItem[]> {
     // Find any Xcode installations on the system
@@ -201,7 +203,7 @@ async function getQuickPickItems(
             return result;
         });
     // Find any Swift toolchains installed via Swiftly
-    const swiftlyToolchains = (await Swiftly.listAvailableToolchains())
+    const swiftlyToolchains = (await Swiftly.listAvailableToolchains(logger))
         .reverse()
         .map<SwiftlyToolchainItem>(toolchainPath => ({
             type: "toolchain",
@@ -297,11 +299,12 @@ async function getQuickPickItems(
  */
 export async function showToolchainSelectionQuickPick(
     activeToolchain: SwiftToolchain | undefined,
+    logger: SwiftLogger,
     cwd?: vscode.Uri
 ) {
     let xcodePaths: string[] = [];
     const selected = await vscode.window.showQuickPick<SelectToolchainItem>(
-        getQuickPickItems(activeToolchain, cwd).then(result => {
+        getQuickPickItems(activeToolchain, logger, cwd).then(result => {
             xcodePaths = result
                 .filter((i): i is XcodeToolchainItem => "category" in i && i.category === "xcode")
                 .map(xcode => xcode.xcodePath);